+++ security/core/src/main/java/org/seedstack/seed/security/internal/realms/AuthenticationTokenWrapper.java
@@ -5,6 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed.security.internal.realms;
 
 import org.apache.shiro.authc.AuthenticationToken;
@@ -33,6 +34,11 @@ public class AuthenticationTokenWrapper implements AuthenticationToken {
     @Override
     public Object getCredentials() {
         return seedToken.getCredentials();
+    }
+
+    @Override
+    public String toString() {
+        return seedToken.name();
     }
 
     /**